nottoomuch-emacs-mailto.pl: add body (if given) in a hook
Message body, if given is now added in (let-copied) temporary version
of message-signature-setup-hook, so that e.g. later-called
message-setup-hook can add it's own content (around the body now
inserted) -- previously (when body was added later), e.g. mml signature
header was added *after* the inserted body.

The message-signature-setup-hook is "liberally abused" to get this done,
not too nice, but I expect it to work in most imaginable cases for years
to come.